# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/meals": {
      "get": {
        "operationId": "MealsController_findAll",
        "parameters": [
          {
            "name": "start",
            "required": true,
            "in": "query",
            "description": "Start date (YYYY-MM-DD)",
            "schema": {
              "example": "2023-12-08",
              "type": "string"
            }
          },
          {
            "name": "end",
            "required": true,
            "in": "query",
            "description": "End date (YYYY-MM-DD)",
            "schema": {
              "example": "2023-12-14",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MealEntity"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "meals"
        ]
      },
      "post": {
        "operationId": "MealsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMealDto"
              },
              "examples": {
                "default": {
                  "summary": "Lunch with Chicken Breast",
                  "value": {
                    "name": "Lunch",
                    "date": "2023-12-08T12:00:00Z",
                    "mealFoods": [
                      {
                        "name": "Chicken Breast",
                        "weight": 150,
                        "calories": 165,
                        "protein": 31,
                        "fats": 3.6,
                        "carbs": 0
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MealEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "meals"
        ]
      }
    },
    "/meals/{id}": {
      "get": {
        "operationId": "MealsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MealEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "meals"
        ]
      },
      "patch": {
        "operationId": "MealsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMealDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MealEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "meals"
        ]
      },
      "delete": {
        "operationId": "MealsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "meals"
        ]
      }
    },
...
    "/meals/{mealId}/foods/{id}": {
      "get": {
        "operationId": "MealFoodsController_findOne",
        "parameters": [
          {
            "name": "mealId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MealFoodEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "meal-foods"
        ]
      },
      "patch": {
        "operationId": "MealFoodsController_update",
        "parameters": [
          {
            "name": "mealId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMealFoodDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MealFoodEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "meal-foods"
        ]
      },
      "delete": {
        "operationId": "MealFoodsController_remove",
        "parameters": [
          {
            "name": "mealId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "meal-foods"
        ]
      }
    },
...
    "/templates/foods": {
      "get": {
        "operationId": "TemplateFoodsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TemplateFoodEntity"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "template-foods"
        ]
      },
      "post": {
        "operationId": "TemplateFoodsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTemplateFoodDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateFoodEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "template-foods"
        ]
      }
    },
    "/templates/foods/{id}": {
      "get": {
        "operationId": "TemplateFoodsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateFoodEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "template-foods"
        ]
      },
      "patch": {
        "operationId": "TemplateFoodsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTemplateFoodDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateFoodEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "template-foods"
        ]
      },
      "delete": {
        "operationId": "TemplateFoodsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "template-foods"
        ]
      }
    },
...
    "/templates/meals/{id}": {
      "get": {
        "operationId": "TemplateMealsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateMealEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "template-meals"
        ]
      },
      "patch": {
        "operationId": "TemplateMealsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTemplateMealDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateMealEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "template-meals"
        ]
      },
      "delete": {
        "operationId": "TemplateMealsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "template-meals"
        ]
      }
    },
...
    "/templates/meals/{templateMealId}/foods/{id}": {
      "get": {
        "operationId": "TemplateMealFoodsController_findOne",
        "parameters": [
          {
            "name": "templateMealId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateMealFoodEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "template-meal-foods"
        ]
      },
      "patch": {
        "operationId": "TemplateMealFoodsController_update",
        "parameters": [
          {
            "name": "templateMealId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTemplateMealFoodDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateMealFoodEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "template-meal-foods"
        ]
      },
      "delete": {
        "operationId": "TemplateMealFoodsController_remove",
        "parameters": [
          {
            "name": "templateMealId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "template-meal-foods"
        ]
      }
    },
...
  },
  "info": {
    "title": "NutriSense API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "MealFoodEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "meal_id": {
            "type": "number",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Chicken Breast"
          },
          "weight": {
            "type": "number",
            "example": 150,
            "description": "Weight in grams"
          },
          "calories": {
            "type": "number",
            "example": 247.5,
            "description": "Calculated calories for the specific weight"
          },
          "protein": {
            "type": "number",
            "example": 46.5,
            "description": "Calculated protein for the specific weight"
          },
          "fats": {
            "type": "number",
            "example": 5.4,
            "description": "Calculated fats for the specific weight"
          },
          "carbs": {
            "type": "number",
            "example": 0,
            "description": "Calculated carbs for the specific weight"
          }
        },
        "required": [
          "id",
          "meal_id",
          "name",
          "weight",
          "calories",
          "protein",
          "fats",
          "carbs"
        ]
      },
      "MealEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "user_id": {
            "type": "number",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Lunch"
          },
          "date": {
            "format": "date-time",
            "type": "string",
            "example": "2023-12-08T12:00:00.000Z"
          },
          "calories": {
            "type": "number",
            "example": 247.5,
            "description": "Sum of meal foods calories"
          },
          "protein": {
            "type": "number",
            "example": 46.5,
            "description": "Sum of meal foods protein"
          },
          "fats": {
            "type": "number",
            "example": 5.4,
            "description": "Sum of meal foods fats"
          },
          "carbs": {
            "type": "number",
            "example": 0,
            "description": "Sum of meal foods carbs"
          },
          "meal_foods": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MealFoodEntity"
            }
          }
        },
        "required": [
          "id",
          "user_id",
          "name",
          "date",
          "calories",
          "protein",
          "fats",
          "carbs",
          "meal_foods"
        ]
      },
      "CreateMealFoodDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Chicken Breast"
          },
          "weight": {
            "type": "number",
            "example": 150,
            "description": "Weight in grams"
          },
          "calories": {
            "type": "number",
            "example": 165,
            "description": "Calories per 100g"
          },
          "protein": {
            "type": "number",
            "example": 31,
            "description": "Protein per 100g"
          },
          "fats": {
            "type": "number",
            "example": 3.6,
            "description": "Fats per 100g"
          },
          "carbs": {
            "type": "number",
            "example": 0,
            "description": "Carbs per 100g"
          }
        },
        "required": [
          "name",
          "weight",
          "calories",
          "protein",
          "fats",
          "carbs"
        ]
      },
      "CreateMealDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Lunch"
          },
          "date": {
            "type": "string",
            "example": "2023-12-08T12:00:00Z"
          },
          "mealFoods": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateMealFoodDto"
            }
          }
        },
        "required": [
          "name",
          "date",
          "mealFoods"
        ]
      },
      "UpdateMealDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Lunch"
          },
          "date": {
            "type": "string",
            "example": "2023-12-08T12:00:00Z"
          },
          "mealFoods": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateMealFoodDto"
            }
          }
        }
      },
      "UpdateMealFoodDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Chicken Breast"
          },
          "weight": {
            "type": "number",
            "example": 150,
            "description": "Weight in grams"
          },
          "calories": {
            "type": "number",
            "example": 165,
            "description": "Calories per 100g"
          },
          "protein": {
            "type": "number",
            "example": 31,
            "description": "Protein per 100g"
          },
          "fats": {
            "type": "number",
            "example": 3.6,
            "description": "Fats per 100g"
          },
          "carbs": {
            "type": "number",
            "example": 0,
            "description": "Carbs per 100g"
          }
        }
      },
...
      "TemplateFoodEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "user_id": {
            "type": "number",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Chicken Breast"
          },
          "calories": {
            "type": "number",
            "example": 165,
            "description": "Per 100g or unit"
          },
          "protein": {
            "type": "number",
            "example": 31,
            "description": "Per 100g or unit"
          },
          "fats": {
            "type": "number",
            "example": 3.6,
            "description": "Per 100g or unit"
          },
          "carbs": {
            "type": "number",
            "example": 0,
            "description": "Per 100g or unit"
          }
        },
        "required": [
          "id",
          "user_id",
          "name",
          "calories",
          "protein",
          "fats",
          "carbs"
        ]
      },
      "CreateTemplateFoodDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Chicken Breast"
          },
          "calories": {
            "type": "number",
            "example": 165,
            "description": "Calories per 100g or unit"
          },
          "protein": {
            "type": "number",
            "example": 31,
            "description": "Protein per 100g or unit"
          },
          "fats": {
            "type": "number",
            "example": 3.6,
            "description": "Fats per 100g or unit"
          },
          "carbs": {
            "type": "number",
            "example": 0,
            "description": "Carbs per 100g or unit"
          }
        },
        "required": [
          "name",
          "calories",
          "protein",
          "fats",
          "carbs"
        ]
      },
      "UpdateTemplateFoodDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Chicken Breast"
          },
          "calories": {
            "type": "number",
            "example": 165,
            "description": "Calories per 100g or unit"
          },
          "protein": {
            "type": "number",
            "example": 31,
            "description": "Protein per 100g or unit"
          },
          "fats": {
            "type": "number",
            "example": 3.6,
            "description": "Fats per 100g or unit"
          },
          "carbs": {
            "type": "number",
            "example": 0,
            "description": "Carbs per 100g or unit"
          }
        }
      },
      "TemplateMealFoodEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "template_meal_id": {
            "type": "number",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Chicken Breast"
          },
          "weight": {
            "type": "number",
            "example": 150,
            "description": "Weight in grams"
          },
          "calories": {
            "type": "number",
            "example": 247.5,
            "description": "Calculated calories for the specific weight"
          },
          "protein": {
            "type": "number",
            "example": 46.5,
            "description": "Calculated protein for the specific weight"
          },
          "fats": {
            "type": "number",
            "example": 5.4,
            "description": "Calculated fats for the specific weight"
          },
          "carbs": {
            "type": "number",
            "example": 0,
            "description": "Calculated carbs for the specific weight"
          }
        },
        "required": [
          "id",
          "template_meal_id",
          "name",
          "weight",
          "calories",
          "protein",
          "fats",
          "carbs"
        ]
      },
      "TemplateMealEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "user_id": {
            "type": "number",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "My Breakfast Template"
          },
          "calories": {
            "type": "number",
            "example": 247.5,
            "description": "Sum of template meal foods"
          },
          "protein": {
            "type": "number",
            "example": 46.5
          },
          "fats": {
            "type": "number",
            "example": 5.4
          },
          "carbs": {
            "type": "number",
            "example": 0
          },
          "template_meal_foods": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TemplateMealFoodEntity"
            }
          }
        },
        "required": [
          "id",
          "user_id",
          "name",
          "calories",
          "protein",
          "fats",
          "carbs",
          "template_meal_foods"
        ]
      },
      "CreateTemplateMealFoodDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Chicken Breast"
          },
          "weight": {
            "type": "number",
            "example": 150,
            "description": "Weight in grams"
          },
          "calories": {
            "type": "number",
            "example": 165,
            "description": "Calories per 100g"
          },
          "protein": {
            "type": "number",
            "example": 31,
            "description": "Protein per 100g"
          },
          "fats": {
            "type": "number",
            "example": 3.6,
            "description": "Fats per 100g"
          },
          "carbs": {
            "type": "number",
            "example": 0,
            "description": "Carbs per 100g"
          }
        },
        "required": [
          "name",
          "weight",
          "calories",
          "protein",
          "fats",
          "carbs"
        ]
      },
...
      "UpdateTemplateMealDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "My Breakfast Template"
          },
          "templateMealFoods": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateTemplateMealFoodDto"
            }
          }
        }
      },
      "UpdateTemplateMealFoodDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Chicken Breast"
          },
          "weight": {
            "type": "number",
            "example": 150,
            "description": "Weight in grams"
          },
          "calories": {
            "type": "number",
            "example": 165,
            "description": "Calories per 100g"
          },
          "protein": {
            "type": "number",
            "example": 31,
            "description": "Protein per 100g"
          },
          "fats": {
            "type": "number",
            "example": 3.6,
            "description": "Fats per 100g"
          },
          "carbs": {
            "type": "number",
            "example": 0,
            "description": "Carbs per 100g"
          }
        }
      },
...
    }
  }
}